New patching method, bugs and fixes
UnityHotSwap no longer uses entire assembly instrumentation to be able
to patch methods. Now, we are forcing methods to be jit-compiled and
replace machine code in memory with jumps and returns to get new code
running. Wowie!

Cleanup and remove a lot of code related to instrumentation and old
patching. This is now more stable then ever! As stable as Unity allows
it to be, before it tries to reload assemblies on its own and fails
miserabely.

Fix some bugs related to IL recompiling. Add support for "constrainted."
opcode. Fix generic method calls but explicitely disallow generic method
emission (DynamicMethod limitation).